Fix issue where stream file is removed before it it passed to the next operation
When using promises the stream file was removed before the promise resolved.
Only cleaning up after the stream is closed solves this problem - and removes the need to sync the delete operation to avoid a race - on close the file will exist so all is well.
Error handling could still be better as currently an error in deleting will crash the app (due to throw) - This needs fixing as it is not a fatal error and really shouldn't do that.